Introduce MongoIndexedSessionRepository.
