+++ src/test/java/org/springframework/session/data/mongo/integration/MongoRepositoryJdkSerializationITest.java
@@ -30,7 +30,7 @@ import org.springframework.session.data.mongo.config.annotation.web.http.EnableM
 import org.springframework.test.context.ContextConfiguration;
 
 /**
- * Integration tests for {@link org.springframework.session.data.mongo.MongoOperationsSessionRepository} that use
+ * Integration tests for {@link org.springframework.session.data.mongo.MongoIndexedSessionRepository} that use
  * {@link JdkMongoSessionConverter} based session serialization.
  *
  * @author Jakub Kubrynski
